[update] openglプロファイル
