+++ BEditor.sln
@@ -84,8 +84,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Core", "src\librari
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Drawing", "src\libraries\BEditor.Drawing\BEditor.Drawing.csproj", "{89EF11AB-317D-4568-BD0E-7A63C8B648B3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Graphics", "src\libraries\BEditor.Graphics\BEditor.Graphics.csproj", "{1FE46B33-0D98-4865-AC04-041B6F969462}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Media", "src\libraries\BEditor.Media\BEditor.Media.csproj", "{09FAE411-8986-49FF-BD78-BF8E183F5340}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Packaging", "src\libraries\BEditor.Packaging\BEditor.Packaging.csproj", "{207E8719-532A-4D76-B89E-D0EB0552B673}"
@@ -95,6 +93,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Settings", "src\libraries\BEditor.Settings\BEditor.Settings.csproj", "{E07ED3D1-0D6C-4696-B756-7ECA627DF953}"
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Shared", "src\libraries\Shared\Shared.shproj", "{865E19ED-CB17-4F80-80BD-5300D135B598}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Graphics", "src\libraries\BEditor.Graphics\BEditor.Graphics.csproj", "{339FD8A4-79C6-4773-A7E3-85D2AADCDD74}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BEditor.Graphics.OpenGL", "src\libraries\BEditor.Graphics.OpenGL\BEditor.Graphics.OpenGL.csproj", "{A9A06B48-9E26-4094-AF1D-057A9D225A39}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
@@ -172,10 +174,6 @@ Global
 		{89EF11AB-317D-4568-BD0E-7A63C8B648B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{89EF11AB-317D-4568-BD0E-7A63C8B648B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{89EF11AB-317D-4568-BD0E-7A63C8B648B3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1FE46B33-0D98-4865-AC04-041B6F969462}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1FE46B33-0D98-4865-AC04-041B6F969462}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1FE46B33-0D98-4865-AC04-041B6F969462}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1FE46B33-0D98-4865-AC04-041B6F969462}.Release|Any CPU.Build.0 = Release|Any CPU
 		{09FAE411-8986-49FF-BD78-BF8E183F5340}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{09FAE411-8986-49FF-BD78-BF8E183F5340}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{09FAE411-8986-49FF-BD78-BF8E183F5340}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -192,6 +190,14 @@ Global
 		{E07ED3D1-0D6C-4696-B756-7ECA627DF953}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E07ED3D1-0D6C-4696-B756-7ECA627DF953}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E07ED3D1-0D6C-4696-B756-7ECA627DF953}.Release|Any CPU.Build.0 = Release|Any CPU
+		{339FD8A4-79C6-4773-A7E3-85D2AADCDD74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{339FD8A4-79C6-4773-A7E3-85D2AADCDD74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{339FD8A4-79C6-4773-A7E3-85D2AADCDD74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{339FD8A4-79C6-4773-A7E3-85D2AADCDD74}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A9A06B48-9E26-4094-AF1D-057A9D225A39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A9A06B48-9E26-4094-AF1D-057A9D225A39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A9A06B48-9E26-4094-AF1D-057A9D225A39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A9A06B48-9E26-4094-AF1D-057A9D225A39}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -215,12 +221,13 @@ Global
 		{80C5AF8F-3209-4FBF-9592-7DF182FC94E0} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{342BBF67-FE7E-4406-8167-FD66BB57B85F} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{89EF11AB-317D-4568-BD0E-7A63C8B648B3} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
-		{1FE46B33-0D98-4865-AC04-041B6F969462} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{09FAE411-8986-49FF-BD78-BF8E183F5340} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{207E8719-532A-4D76-B89E-D0EB0552B673} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{06802D60-FE1C-4C7A-974F-20DCE729BDB5} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{E07ED3D1-0D6C-4696-B756-7ECA627DF953} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 		{865E19ED-CB17-4F80-80BD-5300D135B598} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
+		{339FD8A4-79C6-4773-A7E3-85D2AADCDD74} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
+		{A9A06B48-9E26-4094-AF1D-057A9D225A39} = {C8421BCD-75F3-47F8-85C8-FD47D2807F9C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3B52C4A6-734E-450D-8986-22485F07D198}
