+++ src/libraries/BEditor.Graphics.OpenGL/BEditor.Graphics.OpenGL.csproj
@@ -1,0 +1,34 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <DocumentationFile>BEditor.Graphics.OpenGL.xml</DocumentationFile>
+    </PropertyGroup>
+
+    <Import Project="$(SolutionDir)\props\analyzer.props" />
+
+    <ItemGroup>
+        <PackageReference Include="OpenTK.Graphics" Version="4.6.4" />
+        <PackageReference Include="OpenTK.Windowing.GraphicsLibraryFramework" Version="4.6.4" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\BEditor.Drawing\BEditor.Drawing.csproj" />
+        <ProjectReference Include="..\BEditor.Graphics\BEditor.Graphics.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <Compile Update="Resources\Strings.Designer.cs">
+            <DesignTime>True</DesignTime>
+            <AutoGen>True</AutoGen>
+            <DependentUpon>Strings.resx</DependentUpon>
+        </Compile>
+    </ItemGroup>
+
+    <ItemGroup>
+        <EmbeddedResource Update="Resources\Strings.resx">
+            <Generator>ResXFileCodeGenerator</Generator>
+            <LastGenOutput>Strings.Designer.cs</LastGenOutput>
+        </EmbeddedResource>
+    </ItemGroup>
+
+</Project>
