+++ src/libraries/BEditor.Graphics.OpenGL/Buffers/DepthBuffer.cs
@@ -11,7 +11,7 @@ using System.Threading;
 
 using OpenTK.Graphics.OpenGL4;
 
-namespace BEditor.Graphics
+namespace BEditor.Graphics.OpenGL
 {
     /// <summary>
     /// Represents the depth buffer of the frame buffer.
