+++ src/libraries/BEditor.Graphics.OpenGL/Shaders/LightingShaderFactory.cs
@@ -1,16 +1,16 @@
-﻿// LightingTextureShaderFactory.cs
+﻿// LightingShaderFactory.cs
 //
 // Copyright (C) BEditor
 //
 // This software may be modified and distributed under the terms
 // of the MIT license. See the LICENSE file for details.
 
-namespace BEditor.Graphics
+namespace BEditor.Graphics.OpenGL
 {
-    internal sealed class LightingTextureShaderFactory : ShaderFactory
+    internal sealed class LightingShaderFactory : ShaderFactory
     {
-        private const string TextureFrag =
-            "#version 330\n" +
+        private const string LightFrag =
+            "#version 330 core\n" +
             "struct Material {\n" +
             "    vec4 ambient;\n" +
             "    vec4 diffuse;\n" +
@@ -24,17 +24,16 @@ namespace BEditor.Graphics
             "    vec4 ambient;\n" +
             "    vec4 diffuse;\n" +
             "    vec4 specular;\n" +
-            "};\n" +
+            "};" +
 
             "uniform Light light;\n" +
             "uniform Material material;\n" +
-            "uniform vec4 color;\n" +
-            "uniform sampler2D texture0;\n" +
+
+            "out vec4 FragColor;\n" +
+
             "uniform vec3 viewPos;\n" +
+            "uniform vec4 color;\n" +
 
-            "out vec4 outputColor;\n" +
-
-            "in vec2 texCoord;\n" +
             "in vec3 Normal;\n" +
             "in vec3 FragPos;\n" +
 
@@ -57,36 +56,31 @@ namespace BEditor.Graphics
             "    vec4 specular = light.specular * (spec * material.specular);\n" +
 
             "    vec4 result = ambient + diffuse + specular;\n" +
-            "    outputColor = texture(texture0, texCoord) * result * color;\n" +
+            "    FragColor = result * color;" +
             "}";
 
-        private const string TextureVert =
+        private const string LightVert =
             "#version 330 core\n" +
-
-            "layout(location = 0) in vec3 aPosition;\n" +
-            "layout(location = 1) in vec2 aTexCoord;\n" +
-            "layout(location = 2) in vec3 aNormal;\n" +
+            "layout (location = 0) in vec3 aPos;\n" +
+            "layout (location = 1) in vec3 aNormal;\n" +
 
             "uniform mat4 model;\n" +
             "uniform mat4 view;\n" +
             "uniform mat4 projection;\n" +
 
-            "out vec2 texCoord;\n" +
             "out vec3 Normal;\n" +
             "out vec3 FragPos;\n" +
 
-            "void main(void)\n" +
+            "void main()\n" +
             "{\n" +
-            "    texCoord = aTexCoord;\n" +
-
-            "    gl_Position = vec4(aPosition, 1.0) * model * view * projection;\n" +
-            "    FragPos = vec3(vec4(aPosition, 1.0) * model);\n" +
+            "    gl_Position = vec4(aPos, 1.0) * model * view * projection;\n" +
+            "    FragPos = vec3(vec4(aPos, 1.0) * model);\n" +
             "    Normal = aNormal * mat3(transpose(inverse(model)));\n" +
             "}";
 
         public override Shader Create()
         {
-            return new(TextureVert, TextureFrag);
+            return new(LightVert, LightFrag);
         }
     }
 }
