+++ src/libraries/BEditor.Graphics.OpenGL/Shaders/LineShaderFactory.cs
@@ -5,7 +5,7 @@
 // This software may be modified and distributed under the terms
 // of the MIT license. See the LICENSE file for details.
 
-namespace BEditor.Graphics
+namespace BEditor.Graphics.OpenGL
 {
     internal sealed class LineShaderFactory : ShaderFactory
     {
