+++ src/libraries/BEditor.Graphics.OpenGL/stylecop.json
