+++ src/libraries/BEditor.Graphics/Line.cs
@@ -6,96 +6,65 @@
 // of the MIT license. See the LICENSE file for details.
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
 
-using OpenTK.Graphics.OpenGL4;
+using BEditor.Graphics.Platform;
 
 namespace BEditor.Graphics
 {
     /// <summary>
-    /// Represents an OpenGL line.
+    /// Represents an line.
     /// </summary>
-    public class Line : GraphicsObject
+    public sealed class Line : Drawable
     {
-        private readonly float[] _vertices;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="Line"/> class.
         /// </summary>
         /// <param name="start">The start position of the line.</param>
         /// <param name="end">The end position of the line.</param>
         /// <param name="width">The width of the line.</param>
-        /// <exception cref="GraphicsException">OpenGL error occurred.</exception>
+        /// <exception cref="GraphicsException">Platform is not set.</exception>
         public Line(Vector3 start, Vector3 end, float width)
+            : base(IPlatform.Current.CreateLine(start, end, width))
         {
-            Start = start;
-            End = end;
-            Width = width;
-
-            _vertices = new float[]
-            {
-                start.X, start.Y, start.Z,
-                end.X, end.Y, end.Z,
-            };
-
-            VertexArrayObject = GL.GenVertexArray();
-            VertexBufferObject = GL.GenBuffer();
-            GL.BindVertexArray(VertexArrayObject);
-
-            GL.BindBuffer(BufferTarget.ArrayBuffer, VertexBufferObject);
-            GL.BufferData(BufferTarget.ArrayBuffer, _vertices.Length * sizeof(float), _vertices, BufferUsageHint.StaticDraw);
-
-            GL.VertexAttribPointer(0, 3, VertexAttribPointerType.Float, false, 3 * sizeof(float), IntPtr.Zero);
-            GL.EnableVertexAttribArray(0);
-
-            GL.BindBuffer(BufferTarget.ArrayBuffer, 0);
-            GL.BindVertexArray(0);
-
-            Tool.ThrowGLError();
         }
 
-        /// <inheritdoc/>
-        public override ReadOnlyMemory<float> Vertices => _vertices;
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Line"/> class.
+        /// </summary>
+        /// <param name="impl">The line implementation.</param>
+        public Line(ILineImpl impl)
+            : base(impl)
+        {
+        }
+
+        /// <summary>
+        /// Gets the vertices of this <see cref="Line"/>.
+        /// </summary>
+        public ReadOnlyMemory<float> Vertices => PlatformImpl.Vertices;
 
         /// <summary>
         /// Gets the start position of this <see cref="Line"/>.
         /// </summary>
-        public Vector3 Start { get; }
+        public Vector3 Start => PlatformImpl.Start;
 
         /// <summary>
         /// Gets the end position of this <see cref="Line"/>.
         /// </summary>
-        public Vector3 End { get; }
+        public Vector3 End => PlatformImpl.End;
 
         /// <summary>
         /// Gets the width of this <see cref="Line"/>.
         /// </summary>
-        public float Width { get; }
+        public float Width => PlatformImpl.Width;
 
         /// <summary>
-        /// Gets the VertexBuffer of this <see cref="Line"/>.
+        /// Gets the line implementation.
         /// </summary>
-        public GraphicsHandle VertexBufferObject { get; }
-
-        /// <summary>
-        /// Gets the VertexArray of this <see cref="Line"/>.
-        /// </summary>
-        public GraphicsHandle VertexArrayObject { get; }
-
-        /// <inheritdoc/>
-        public override void Draw()
-        {
-            GL.LineWidth(Width);
-
-            GL.BindVertexArray(VertexArrayObject);
-            GL.DrawArrays(PrimitiveType.Lines, 0, 2);
-        }
-
-        /// <inheritdoc/>
-        protected override void Dispose(bool disposing)
-        {
-            GL.DeleteVertexArray(VertexArrayObject);
-            GL.DeleteBuffer(VertexBufferObject);
-        }
+        public new ILineImpl PlatformImpl => (ILineImpl)base.PlatformImpl;
     }
 }
